add popular controller

+++ MyCinemaApp/src/main/java/com/example/MyCinemaApp/entity/Genre.java
@@ -19,11 +19,12 @@ public class Genre {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @Column(columnDefinition = "VARCHAR(255) NOT NULL")
     private String name;
 
     @ManyToMany(mappedBy = "genreSet")
     private Set<Cinema> cinemaSet;
+
 }
